#265 minor corrections, rearrange and improve tests
